작업내용
1. 컨트롤러 기능 정리, 수정
