+++ PandaClimb/src/main/resources/templates/user/edit.html
@@ -14,7 +14,7 @@
 	<div>
 	<form action = "edit.do" method = "post">
 		
-	<input type = "text" name = "userId"  th:value="${userDto.userId}">
+	<input type = "hidden" name = "userId"  th:value="${userDto.userId}">
 	
 	<div class="mt-4">
 		<div class="col-lg-4 offset-lg-4 col-md-6 offset-md-3 col-sm-8 offset-sm-2">
